Adding  -All:$true to cmdlet
Get-AzureADServicePrincipal -All:$true 
Adding -All:$true to Get-AzureADServicePrincipal allows the cmdlet to look through all the Service Principle accounts.  I when through a lot of work to figure out that the account was there but was not showing up in the first 100 in the query that is allowed without the -All:$true.
There are two places on this page the change needs to be added.